requirements.txt
picamera2>=0.3.34
flask>=3.1.0
Pillow>=11.0.0
numpy>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
anthropic>=0.40.0
